using QtSerialPort
exceptions used in low-level library are not correctly fetched by Qt code...

+++ linux/lib/GloveWidget.h
@@ -74,7 +74,7 @@ private:
 		QString  name;
 		QDomNode styleNode;
 		QString  styleString;
-		unsigned short iFillStart;
+		int      iFillStart;
 		short channel; // -1 if unassigned
 	};
